requirements.txt
EasyProcess==0.2.2
EbookLib==0.15
Pillow==3.2.0
PyVirtualDisplay==0.2
SpeechRecognition==3.4.3
XlsxWriter==0.8.4
argcomplete==1.1.1
argparse==1.2.1
beautifulsoup4==4.4.1
chardet==2.3.0
lxml==3.6.0
pdfminer==20140328
python-docx==0.8.5
python-pptx==0.5.8
selenium==2.53.1
six==1.10.0
textract==1.4.0
wsgiref==0.1.2
xlrd==0.9.4